--- py/orbit/danilov_envelope/__init__.py
@@ -1,8 +1,0 @@
-from .danilov_envelope_20 import DanilovEnvelope20
-from .danilov_envelope_20 import DanilovEnvelopeTracker20
-from .danilov_envelope_22 import DanilovEnvelope22
-from .danilov_envelope_22 import DanilovEnvelopeTracker22
-from .danilov_envelope_solver_nodes import DanilovEnvelopeSolverNode20
-from .danilov_envelope_solver_nodes import DanilovEnvelopeSolverNode22
-from .danilov_envelope_solver_lattice_modifications import add_danilov_envelope_solver_nodes_20
-from .danilov_envelope_solver_lattice_modifications import add_danilov_envelope_solver_nodes_22
