--- py/orbit/danilov_envelope/meson.build
@@ -1,14 +1,0 @@
-py_sources = files([
-    '__init__.py',
-    'danilov_envelope_20.py',
-    'danilov_envelope_22.py',
-    'danilov_envelope_solver_nodes.py',
-    'danilov_envelope_solver_lattice_modifications.py',
-    'transfer_matrix.py',
-    'utils.py',
-])
-
-python.install_sources(
-    py_sources,
-    subdir: 'orbit/danilov_envelope',
-)
