+++ py/orbit/envelope/meson.build
@@ -1,0 +1,8 @@
+py_sources = files([
+    '__init__.py',
+])
+
+python.install_sources(
+    py_sources,
+    subdir: 'orbit/envelope',
+)
