+++ py/orbit/envelope/utils.py
